% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Auto Correlation.R
\name{acfSamples.iSIR}
\alias{acfSamples.iSIR}
\title{Method for iSIR class, only plots for Beta, Gamma and newI.}
\usage{
\method{acfSamples}{iSIR}(samples, plot = TRUE, timePoints = NA)
}
\description{
Method for iSIR class, only plots for Beta, Gamma and newI.
}
