% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Auto Correlation.R
\name{acfSamples}
\alias{acfSamples}
\title{A function that plots and prints ACF plots and effective sample size for samples from the
metropolis hastings algorithm using the acf() function.}
\usage{
acfSamples(samples, plot, timePoints)
}
\description{
A function that plots and prints ACF plots and effective sample size for samples from the
metropolis hastings algorithm using the acf() function.
}
