% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Auto Correlation.R
\name{acfSamples.rSIR}
\alias{acfSamples.rSIR}
\title{Method for rSIR class, only plots for Beta, Gamma and newI.}
\usage{
\method{acfSamples}{rSIR}(samples, plot = TRUE, timePoints = NA)
}
\description{
Method for rSIR class, only plots for Beta, Gamma and newI.
}
