% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Auto Correlation.R
\name{acfSamples.SIR}
\alias{acfSamples.SIR}
\title{Method for SIR class, only plots for Beta and Gamma.}
\usage{
\method{acfSamples}{SIR}(samples, plot = TRUE)
}
\description{
Method for SIR class, only plots for Beta and Gamma.
}
