% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Data Generation.R
\name{dataGen}
\alias{dataGen}
\title{Generates population data from given parameters}
\usage{
dataGen(epiModel)
}
\arguments{
\item{epiModel}{epidemic model in the format of a specified class}
}
\value{
object in the same format as the given object
}
\description{
Provided values of population sizes but not changes at each time step
this function calculates the changes and vice-versa. It will add these
values to a class of format provided. The purpose is that if we already
calculate these values here we won't have to reculate them everytime
we wan to find the likelihood etc. Currently on takes objects of the
SIR class.
}
