% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Data Generation.R
\name{dataGen.SIR}
\alias{dataGen.SIR}
\title{Generates population data from given parameters
for the SIR class.}
\usage{
\method{dataGen}{SIR}(epiModel)
}
\description{
Generates population data from given parameters
for the SIR class.
}
