% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{independenceSampler.rSIR}
\alias{independenceSampler.rSIR}
\title{Independence sampler for the rSIR class.
First generates beta and gamma from a normal distribution centred around the mean
value of their priors. Then uses those values to generate a newI or newR using
the standard Binomial distribution from the SIR model.}
\usage{
\method{independenceSampler}{rSIR}(epiModel, hyperParameters)
}
\description{
Independence sampler for the rSIR class.
First generates beta and gamma from a normal distribution centred around the mean
value of their priors. Then uses those values to generate a newI or newR using
the standard Binomial distribution from the SIR model.
}
