% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{independenceSampler.SIR}
\alias{independenceSampler.SIR}
\title{Independence sampler for the SIR class.
First generates beta and gamma from a normal distribution centred around the mean
value of their priors.}
\usage{
\method{independenceSampler}{SIR}(epiModel, hyperParameters, i)
}
\description{
Independence sampler for the SIR class.
First generates beta and gamma from a normal distribution centred around the mean
value of their priors.
}
