% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Initial Values.R
\name{initialValues.iSIR}
\alias{initialValues.iSIR}
\title{Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters and then
generates newI.
from it.}
\usage{
\method{initialValues}{iSIR}(epiModel, hyperParameters)
}
\description{
Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters and then
generates newI.
from it.
}
