% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Initial Values.R
\name{initialValues}
\alias{initialValues}
\title{function that sets up initial values for MH algorithm.}
\usage{
initialValues(epiModel, hyperParameters)
}
\description{
function that sets up initial values for MH algorithm.
}
