% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Initial Values.R
\name{initialValues.rSIR}
\alias{initialValues.rSIR}
\title{Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters and then
generates newR.
from it.}
\usage{
\method{initialValues}{rSIR}(epiModel, hyperParameters)
}
\description{
Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters and then
generates newR.
from it.
}
