% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Initial Values.R
\name{initialValues.SIR}
\alias{initialValues.SIR}
\title{Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters.
from it.}
\usage{
\method{initialValues}{SIR}(epiModel, hyperParameters)
}
\description{
Intial values for SIR model. Just sets Beta and Gamma to
the start value specified in the hyperparameters.
from it.
}
