% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{logIndependenceSampler.iSIR}
\alias{logIndependenceSampler.iSIR}
\title{Method for the iSIR class.
finds the log density for the values of Beta and Gamma being
generated and then that the binomial chain generated the values of
newI.}
\usage{
\method{logIndependenceSampler}{iSIR}(epiModel, hyperParameters)
}
\description{
Method for the iSIR class.
finds the log density for the values of Beta and Gamma being
generated and then that the binomial chain generated the values of
newI.
}
