% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{logIndependenceSampler}
\alias{logIndependenceSampler}
\title{Finds the log probability of that the indepdence sampler genreated
that sample.}
\usage{
logIndependenceSampler(epiModel, hyperParameters)
}
\description{
Finds the log probability of that the indepdence sampler genreated
that sample.
}
