% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{logIndependenceSampler.rSIR}
\alias{logIndependenceSampler.rSIR}
\title{Method for the iSIR class.
finds the log density for the values of Beta and Gamma being
generated and then that the binomial chain generated the values of
newR.}
\usage{
\method{logIndependenceSampler}{rSIR}(epiModel, hyperParameters)
}
\description{
Method for the iSIR class.
finds the log density for the values of Beta and Gamma being
generated and then that the binomial chain generated the values of
newR.
}
