% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{logIndependenceSampler.SIR}
\alias{logIndependenceSampler.SIR}
\title{Method for the SIR class.
finds the log density for the values of Beta and Gamma being
generated.}
\usage{
\method{logIndependenceSampler}{SIR}(epiModel, hyperParameters)
}
\description{
Method for the SIR class.
finds the log density for the values of Beta and Gamma being
generated.
}
