% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-log Likelihood.R
\name{logLikelihood}
\alias{logLikelihood}
\title{Log-Likelihood}
\usage{
logLikelihood(epiModel)
}
\arguments{
\item{epiModel}{An epidemic model with know parameters}
}
\value{
the log-likelihood
}
\description{
Given a model with no unknown parameters this calculates the natural logarithm
of the likelihood. Only accepts SIR models.
}
