% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Prior Probability.R
\name{logPrior.iSIR}
\alias{logPrior.iSIR}
\title{Log prior for when newI is a parameter in the SIR model.
The priors from Beta and Gamma are considered to be Gamma(alpha, beta).
The prior for newI(t) is the discrete Unif(max(newR(t+1) - I(t), 0), S(t)).}
\usage{
\method{logPrior}{iSIR}(epiModel, priors)
}
\description{
Log prior for when newI is a parameter in the SIR model.
The priors from Beta and Gamma are considered to be Gamma(alpha, beta).
The prior for newI(t) is the discrete Unif(max(newR(t+1) - I(t), 0), S(t)).
}
