% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Prior Probability.R
\name{logPrior}
\alias{logPrior}
\title{Calculates the log of the prior probability for the model}
\usage{
logPrior(epiModel, priors)
}
\arguments{
\item{epiModel}{some epidemic model with no missing data}

\item{priors}{the hyperparameters of the priors in list form}
}
\value{
the value of the log prior probability
}
\description{
Calculates the log of the prior probability for the model
}
