% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Prior Probability.R
\name{logPrior.rSIR}
\alias{logPrior.rSIR}
\title{Log prior for when newR is a parameter in the SIR model.
The priors from Beta and Gamma are considered to be Gamma(alpha, beta).
The prior for newR(t) is the discrete Unif(0, I(t)).
Note, I am sure the prior for newR(t) can be more infromative, e.g.
needing atleast one infectious person until the end etc. I've not
gotten round to thinking about it yet, if I ever do.}
\usage{
\method{logPrior}{rSIR}(epiModel, priors)
}
\description{
Log prior for when newR is a parameter in the SIR model.
The priors from Beta and Gamma are considered to be Gamma(alpha, beta).
The prior for newR(t) is the discrete Unif(0, I(t)).
Note, I am sure the prior for newR(t) can be more infromative, e.g.
needing atleast one infectious person until the end etc. I've not
gotten round to thinking about it yet, if I ever do.
}
