% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Prior Probability.R
\name{logPrior.SIR}
\alias{logPrior.SIR}
\title{Log prior for when Beta and Gamma are the paramters considered.}
\usage{
\method{logPrior}{SIR}(epiModel, priors)
}
\description{
Log prior for when Beta and Gamma are the paramters considered.
}
