% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Proposal Probability Master Function.R
\name{logProposal}
\alias{logProposal}
\title{Master function for find the log proposals for each proposal function.
Sends the models and revelant parameters to the relevant function for the
proposal method given.}
\usage{
logProposal(epiModel1, epiModel2, proposal, hyperParameters)
}
\description{
Master function for find the log proposals for each proposal function.
Sends the models and revelant parameters to the relevant function for the
proposal method given.
}
