% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{logRandomWalk.iSIR}
\alias{logRandomWalk.iSIR}
\title{Method for the iSIR class.
Doesn't find the density for Beta, Gamma since these will cancel
in the acceptance probability. Sums the log densities for newI if
Beta and Gamma came from the old model.}
\usage{
\method{logRandomWalk}{iSIR}(epiModel1, epiModel2)
}
\description{
Method for the iSIR class.
Doesn't find the density for Beta, Gamma since these will cancel
in the acceptance probability. Sums the log densities for newI if
Beta and Gamma came from the old model.
}
