% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{logRandomWalk}
\alias{logRandomWalk}
\title{Finds the log probability that the random walk generated the given candidate.}
\usage{
logRandomWalk(epiModel1, epiModel2)
}
\description{
Finds the log probability that the random walk generated the given candidate.
}
