% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{logRandomWalk.SIR}
\alias{logRandomWalk.SIR}
\title{Method for the SIR class.
Just returns 0 since its symmetric.}
\usage{
\method{logRandomWalk}{SIR}(epiModel1, epiModel2)
}
\description{
Method for the SIR class.
Just returns 0 since its symmetric.
}
