% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Metropolis Hastings.R
\name{metropolisHastings}
\alias{metropolisHastings}
\title{Generates the misisng values/estimates parameters using the
Metropolis Hastings algorithm. Returns a list of samples from
the posterior distribution. Currently accepts iSIR or rSIR objects
and proposal takes randomWalk or independenceSampler}
\usage{
metropolisHastings(
  epiModel,
  proposal,
  hyperParameters,
  samples = 1000,
  burnin = 1000,
  keep = 100
)
}
\arguments{
\item{epiModel}{Some object with parameters/missing data to impute}

\item{proposal}{the proposal function}

\item{hyperParameters}{the hyperparameters for the priors, in list form + variance for randomwalk}

\item{samples}{the target number of samples to be left with}

\item{burnin}{the amount of samples to be dropped as burnin}

\item{keep}{the algorithm will keep 1 in keep of the generated samples}
}
\description{
Generates the misisng values/estimates parameters using the
Metropolis Hastings algorithm. Returns a list of samples from
the posterior distribution. Currently accepts iSIR or rSIR objects
and proposal takes randomWalk or independenceSampler
}
