% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-Maximum Likelihood Esimator.R
\name{MLE}
\alias{MLE}
\title{Maximum Likelihood Estimate}
\usage{
MLE(epiModel, bounds)
}
\arguments{
\item{epiModel}{An epidemic model}

\item{bounds}{list of vectors of lower and upper bounds for each parameter to be estimated}
}
\value{
the MLE
}
\description{
Uses optim to make an MLE for the parameters of given epidemic model.
Uses L-BFGS-B so bounds can be provided as a list of vectors.
Currently accepts only SIR class models
}
