% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle-Maximum Likelihood Esimator.R
\name{MLE.SIR}
\alias{MLE.SIR}
\title{MLE for SIR class}
\usage{
\method{MLE}{SIR}(epiModel, bounds = list(Beta = c(10^-10, 0.1), Gamma = c(10^-10, 0.1)))
}
\description{
MLE for SIR class
}
