% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Plot Methods.R
\name{plot.SIR}
\alias{plot.SIR}
\title{Plots some useful graphs of the SIR Model}
\usage{
\method{plot}{SIR}(epiModel)
}
\arguments{
\item{epiModel}{An epidemic model of SIR class}
}
\value{
ggplot object
}
\description{
Plots the populations over time
Requires ggplot
}
