% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Plot Methods.R
\name{plotSamples.iSIR}
\alias{plotSamples.iSIR}
\title{Plots some useful graphs of the bayesian-iSIR Model}
\usage{
\method{plotSamples}{iSIR}(x, timePoints = NA)
}
\arguments{
\item{samples}{An object containing many epidemic model of SIR class}
}
\value{
ggplot object
}
\description{
Plots the populations over time
Requires ggplot
}
