% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Plot Methods.R
\name{plotSamples.rSIR}
\alias{plotSamples.rSIR}
\title{Plots some useful graphs of the bayesian-rSIR Model}
\usage{
\method{plotSamples}{rSIR}(x, timePoints = NA)
}
\arguments{
\item{samples}{An object containing many epidemic model of SIR class}
}
\value{
ggplot object
}
\description{
Plots the populations over time
Requires ggplot
}
