% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Plot Methods.R
\name{plotSamples.SIR}
\alias{plotSamples.SIR}
\title{Plots some useful graphs of the bayesian-SIR Model}
\usage{
\method{plotSamples}{SIR}(x, timePoints)
}
\arguments{
\item{samples}{An object containing many epidemic model of SIR class}
}
\value{
ggplot object
}
\description{
Plots the populations over time
Requires ggplot
}
