% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Probability Function.R
\name{probGen}
\alias{probGen}
\title{Probability Generator}
\usage{
probGen(x)
}
\arguments{
\item{x}{some non-negative value}
}
\value{
a value on the range (0,1)
}
\description{
Converts a given value into a value in the region (0,1).
Uses the exponential cdf. This funtion is only defined
seperately so that it can be easily changed if needed.
}
