% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{randomWalk.rSIR}
\alias{randomWalk.rSIR}
\title{Generates a new candidate using the randomwalk for rSIR class.
Generates Beta, Gamma from normal centered around the past value of
Beta, Gamma and with standard deviation given in hyperparameters.}
\usage{
\method{randomWalk}{rSIR}(epiModel, hyperParameters)
}
\description{
Generates a new candidate using the randomwalk for rSIR class.
Generates Beta, Gamma from normal centered around the past value of
Beta, Gamma and with standard deviation given in hyperparameters.
}
