% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{randomWalk.SIR}
\alias{randomWalk.SIR}
\title{Generates a new candidate using the randomwalk for SIR class.
Generates Beta, Gamma from normal centered around the past value of
Beta, Gamma and with standard deviation given in hyperparameters.}
\usage{
\method{randomWalk}{SIR}(epiModel, hyperParameters)
}
\description{
Generates a new candidate using the randomwalk for SIR class.
Generates Beta, Gamma from normal centered around the past value of
Beta, Gamma and with standard deviation given in hyperparameters.
}
