% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Simulating Data.R
\name{simulate.rSIR}
\alias{simulate.rSIR}
\title{Generates Simulated Epidemics for rSIR}
\usage{
\method{simulate}{rSIR}(epiModel)
}
\arguments{
\item{epiModel}{An epidemic model with know parameters}
}
\value{
an object containing the generated data
}
\description{
Given an model with specified parameters, e.g. Beta, Gamma and population size,
R and newR this function returns a stochastically simulated epidemic.
}
