% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Simulating Data.R
\name{simulate.SIR}
\alias{simulate.SIR}
\title{Generates Simulated Epidemics}
\usage{
\method{simulate}{SIR}(epiModel, t.max = 1000)
}
\arguments{
\item{epiModel}{An epidemic model with know parameters}

\item{t.max}{the largest number of time-steps to be taken}
}
\value{
an object containing the generated data
}
\description{
Given an model with specified parameters, e.g. Beta, Gamma and population size,
this function returns a stochastically simulated epidemic. The model used will
be of the same type as the provided model. Currently only accepts SIR models.
}
