% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Simulating Data.R
\name{simulateDataSet}
\alias{simulateDataSet}
\title{Generates Many Simulated Epidemics}
\usage{
simulateDataSet(epiModel, N, t.max)
}
\arguments{
\item{epiModel}{An epidemic model with know parameters}

\item{N}{number of epidemics to simulate}

\item{t.max}{the largest number of time-steps to be taken}
}
\value{
a list containing two lists of the generated data
}
\description{
Given an model with specified parameters, e.g. Beta, Gamma and population size,
this function returns 'n' stochastically simulated epidemics. Will split these into
testing data, with only the number of new infections, N and the timestep and into
the complete data. Currently only accepts SIR models.
}
