% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-Simulating Data.R
\name{simulateDataSet.SIR}
\alias{simulateDataSet.SIR}
\title{Generates Many Simulated Epidemics for the SIR class}
\usage{
\method{simulateDataSet}{SIR}(epiModel, N, t.max = 1000, seed = NA)
}
\description{
Generates Many Simulated Epidemics for the SIR class
}
