% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-SIR Class Definitions.R
\name{SIR}
\alias{SIR}
\title{Creates an object of class SIR, iSIR or rSIR}
\usage{
SIR(
  S = NULL,
  I = NULL,
  R = NULL,
  newI = NULL,
  newR = NULL,
  N = NULL,
  Beta = NULL,
  Gamma = NULL,
  t.step = 1
)
}
\arguments{
\item{S}{Time series of the Susceptible Population}

\item{I}{Time series of the Infectious Population}

\item{R}{Time series of the Recovered Population}

\item{newI}{Time series of new infections}

\item{newR}{Time series of new recoveries}

\item{N}{The size of the population}

\item{Beta}{The infectious weight of one person in I}

\item{Gamma}{The recovery rate}

\item{t.step}{The amount of time in each step of the model, actual value depends on the units of Beta and Gamma}
}
\value{
Object of class SIR
}
\description{
Takes any combination of the values in the SIR model and uses dataGen to generate
unspecified data where it can. If S, I and newI but R and newR are then it will
generate an object of class iSIR that inherits from the SIR class. This class functions
the same for most functions other than in the metropolis hasting algorithm where it treats
newI as a parameter to generate. if R, I and newR are missing then it will make an object of
class rSIR which does the same but treats newR as a parameter.
}
