% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Summary Methods.R
\name{summarise.iSIR}
\alias{summarise.iSIR}
\title{Method that returns a epidemic model with average values of its
parameters, calculated from the samples, for the iSIR class}
\usage{
\method{summarise}{iSIR}(samples, average)
}
\description{
Method that returns a epidemic model with average values of its
parameters, calculated from the samples, for the iSIR class
}
