% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Summary Methods.R
\name{summarise}
\alias{summarise}
\title{A generic function to return a summary model using either the mean, median or mode of the estimated paratmers
Unused at the moment.}
\usage{
summarise(samples, average = mean)
}
\description{
A generic function to return a summary model using either the mean, median or mode of the estimated paratmers
Unused at the moment.
}
