% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Summary Methods.R
\name{summarise.rSIR}
\alias{summarise.rSIR}
\title{Method that returns a epidemic model with mean values of its
parameters, calculated from the samples, for the rSIR class}
\usage{
\method{summarise}{rSIR}(samples, average)
}
\description{
Method that returns a epidemic model with mean values of its
parameters, calculated from the samples, for the rSIR class
}
