% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation-Summary Methods.R
\name{summarise.SIR}
\alias{summarise.SIR}
\title{Method that returns a epidemic model with mean values of its
parameters, calculated from the samples, for the SIR class}
\usage{
\method{summarise}{SIR}(samples, average)
}
\description{
Method that returns a epidemic model with mean values of its
parameters, calculated from the samples, for the SIR class
}
