% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Independence Sampler.R
\name{independenceSampler}
\alias{independenceSampler}
\title{Generates a candidate sample from a predetermined fixed distribution}
\usage{
independenceSampler(epiModel, hyperParameters, i)
}
\arguments{
\item{epiModel}{epidemic model}

\item{hyperParameters}{list of hyperparameters, used in distribution}
}
\description{
Generates a candidate sample from a predetermined fixed distribution
}
