% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc-Random Walk.R
\name{randomWalk}
\alias{randomWalk}
\title{Generates a new candidate using the randomwalk method.}
\usage{
randomWalk(epiModel, priors)
}
\description{
Generates a new candidate using the randomwalk method.
}
